Use components instead of config classes
